Refactor backend package to com.ht0328.uma_ai_selecter and remove demo code
- Rename package com.example.demo -> com.ht0328.uma_ai_selecter
- Delete News and WireMock verification features
- Update build.gradle.kts and settings.gradle.kts
- Clean up openapi.yml and application.yml
