--- api/src/main/kotlin/com/example/demo/news/entity/NewsItem.kt
@@ -1,11 +1,0 @@
-package com.example.demo.news.entity
-
-import java.time.LocalDateTime
-
-data class NewsItem(
-    val id: Long? = null,
-    val title: String,
-    val content: String,
-    val publishedAt: LocalDateTime,
-    val createdAt: LocalDateTime = LocalDateTime.now()
-)
