+++ api/src/main/resources/application.yml
@@ -14,10 +14,6 @@ spring:
       hibernate:
         dialect: org.hibernate.dialect.PostgreSQLDialect
 
-app:
-  wiremock:
-    base-url: ${WIREMOCK_BASE_URL:http://wiremock:8080}
-
 management:
   endpoints:
     web:
