--- api/src/main/resources/openapi/news.yml
@@ -1,51 +1,0 @@
-openapi: 3.0.3
-info:
-  title: News API
-  version: 1.0.0
-paths:
-  /api/news:
-    get:
-      tags:
-        - News
-      summary: Get all news items
-      operationId: getNews
-      responses:
-        '200':
-          description: Successful response
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/NewsItem'
-  /api/news/sync:
-    post:
-      tags:
-        - News
-      summary: Sync news from external source
-      operationId: syncNews
-      responses:
-        '200':
-          description: Sync successful
-components:
-  schemas:
-    NewsItem:
-      type: object
-      properties:
-        id:
-          type: integer
-          format: int64
-        title:
-          type: string
-        content:
-          type: string
-        publishedAt:
-          type: string
-          format: date-time
-        createdAt:
-          type: string
-          format: date-time
-      required:
-        - title
-        - content
-        - publishedAt
