+++ api/src/test/kotlin/com/ht0328/uma_ai_selecter/UmaAiSelecterApplicationTests.kt
@@ -1,10 +1,10 @@
-package com.example.demo
+package com.ht0328.uma_ai_selecter
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class DemoApplicationTests {
+class UmaAiSelecterApplicationTests {
 
 	@Test
 	fun contextLoads() {
